+++ rest_framework_json_api/serializers.py
@@ -1,5 +1,5 @@
 import inflection
-import six
+from django.core.exceptions import ObjectDoesNotExist
 from django.db.models.query import QuerySet
 from django.utils.translation import ugettext_lazy as _
 from rest_framework.exceptions import ParseError
@@ -257,8 +257,7 @@ class PolymorphicSerializerMetaclass(SerializerMetaclass):
         return new_class
 
 
-@six.add_metaclass(PolymorphicSerializerMetaclass)
-class PolymorphicModelSerializer(ModelSerializer):
+class PolymorphicModelSerializer(ModelSerializer, metaclass=PolymorphicSerializerMetaclass):
     """
     A serializer for polymorphic models.
     Useful for "lazy" parent models. Leaves should be represented with a regular serializer.